Daily Changes: Feb 1, 2019
Summary: We got paychecks from Bridgewater and Chick-Fil-A. This was Rebecca's
last full Chick-Fil-A paycheck, she recieved $495 and we had $200 budgeted, that's
a surplus of $295. That extra is going to go into savings as we don't have a
need for it.

Financial Mood:
- Rebecca: I don't love our situation right now, I think we need more
savings and more income. Though, I think we have finally taken the
steps to get to where I want us to go. I'm happy with our rate of
progress but not with our current level.
- Jacob: I think we're improving at a good rate, I'm worried about the
state of our renewable cushion as the surplus income we have month to
month is very small. I don't like the current state of our cashflow. I
am satisfied with the rate of change and our trajectory however, our
current level makes me uncomfortable.

Biggest Hurdel:
- Lifestyle creep-- as we pay off Ryan and Rebecca gets a new job it
will be easy to get complpacent and sit on or laurels.

Short-Term Goals:
- Get Ryan paid off
- Remove small bills like Sprint and Prime Storage
- Reduce our rent as much as possible

Long-Term Goals:
- Find a way to have no "shelter" payment. Neigher rent nor a mortgague.